+++ openmls/src/group/public_group/staged_commit.rs
@@ -123,6 +123,9 @@ impl PublicGroup {
                 FromCommittedProposalsError::LibraryError(e) => StageCommitError::LibraryError(e),
                 FromCommittedProposalsError::ProposalNotFound => StageCommitError::MissingProposal,
                 FromCommittedProposalsError::SelfRemoval => StageCommitError::AttemptedSelfRemoval,
+                FromCommittedProposalsError::DuplicatePskId(psk_id) => {
+                    StageCommitError::DuplicatePskId(psk_id)
+                }
             }
         })?;